+++ tests/Feature/SwarmServiceProviderTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use BuiltByBerry\LaravelSwarm\Commands\MakeSwarmCommand;
 use BuiltByBerry\LaravelSwarm\Commands\SwarmCancelCommand;
+use BuiltByBerry\LaravelSwarm\Commands\SwarmHealthCommand;
 use BuiltByBerry\LaravelSwarm\Commands\SwarmHistoryCommand;
 use BuiltByBerry\LaravelSwarm\Commands\SwarmPauseCommand;
 use BuiltByBerry\LaravelSwarm\Commands\SwarmPruneCommand;
@@ -54,6 +55,7 @@ test('the make swarm command is registered', function () {
 
     expect(array_key_exists('make:swarm', $commands))->toBeTrue();
     expect($commands['make:swarm'])->toBeInstanceOf(MakeSwarmCommand::class);
+    expect($commands['swarm:health'])->toBeInstanceOf(SwarmHealthCommand::class);
     expect($commands['swarm:prune'])->toBeInstanceOf(SwarmPruneCommand::class);
     expect($commands['swarm:status'])->toBeInstanceOf(SwarmStatusCommand::class);
     expect($commands['swarm:history'])->toBeInstanceOf(SwarmHistoryCommand::class);